+++ cypress/support/index.js
@@ -39,9 +39,11 @@ Cypress.Commands.overwrite(
   (originalFn, subject, string, options) => originalFn(
     subject,
     string,
-    Object.assign({}, options, { delay: 100 }),
+    Object.assign({}, options, { delay: 75 }),
   ),
 );
+
+Cypress.Commands.overwrite('log', (subject, message) => cy.task('log', message));
 
 Cypress.Commands.add('radioButtonComponent', (radioButtonLabel) => {
   radioButtonComponent().each(($el) => {
@@ -78,9 +80,5 @@ function getItem(selector, counter) {
 }
 
 Cypress.Commands.add('iFrame', (selector) => { getItem(selector, 50); });
-
-before(() => {
-  cy.wait(1000, { log: DEBUG_FLAG });
-});
 
 Cypress.Screenshot.defaults({ screenshotOnRunFailure: DEBUG_FLAG });