+++ cypress/support/step-definitions/icon-steps.js
@@ -9,8 +9,8 @@ Then('Icon height is set to {string}', (height) => {
 });
 
 Then('radius is set to {string}', (radius) => {
-  icon().should('have.css', 'border-bottom-left-radius', radius);
-  icon().should('have.css', 'border-bottom-right-radius', radius);
-  icon().should('have.css', 'border-top-left-radius', radius);
-  icon().should('have.css', 'border-top-right-radius', radius);
+  icon().should('have.css', 'border-bottom-left-radius', radius)
+    .and('have.css', 'border-bottom-right-radius', radius)
+    .and('have.css', 'border-top-left-radius', radius)
+    .and('have.css', 'border-top-right-radius', radius);
 });