+++ cypress/support/step-definitions/textbox-steps.js
@@ -6,18 +6,17 @@ import {
   fieldHelpPreviewByPosition, tooltipPreviewByPosition, labelNoIFrame, commonDataElementInputPreviewNoIframe,
 } from '../../locators';
 import { DEBUG_FLAG } from '..';
+import { positionOfElement } from '../helper';
 
 const TEXT_ALIGN = 'text-align';
-const FIRST_ELEMENT = 0;
-const SECOND_ELEMENT = 1;
 
 Then('Textbox placeholder is set to {string}', (text) => {
   textbox().children().should('have.attr', 'placeholder', text);
 });
 
 Then('Multiple Textbox placeholder is set to {string}', (text) => {
-  textbox(FIRST_ELEMENT).children().should('have.attr', 'placeholder', text);
-  textbox(SECOND_ELEMENT).children().should('have.attr', 'placeholder', text);
+  textbox(positionOfElement('first')).children().should('have.attr', 'placeholder', text);
+  textbox(positionOfElement('second')).children().should('have.attr', 'placeholder', text);
 });
 
 Then('Textbox component is disabled', () => {
@@ -28,10 +27,10 @@ Then('Textbox component is disabled', () => {
 
 Then('Textbox multiple component is disabled', () => {
   cy.wait(100, { log: DEBUG_FLAG }); // added due to animation changing
-  textbox(FIRST_ELEMENT).children()
+  textbox(positionOfElement('first')).children()
     .should('have.css', 'color', 'rgba(0, 0, 0, 0.55)')
     .and('have.css', 'cursor', 'not-allowed');
-  textbox(SECOND_ELEMENT).children()
+  textbox(positionOfElement('second')).children()
     .should('have.css', 'color', 'rgba(0, 0, 0, 0.55)')
     .and('have.css', 'cursor', 'not-allowed');
 });
@@ -46,37 +45,40 @@ Then('Textbox component is not disabled', () => {
 
 Then('Textbox multiple component is not disabled', () => {
   cy.wait(100, { log: DEBUG_FLAG }); // added due to animation changing
-  textbox(FIRST_ELEMENT).should('not.be.disabled');
-  textbox(FIRST_ELEMENT).children()
+  textbox(positionOfElement('first')).should('not.be.disabled');
+  textbox(positionOfElement('first')).children()
     .should('not.have.css', 'color', 'rgba(0, 0, 0, 0.55)')
     .and('not.have.css', 'cursor', 'not-allowed');
-  textbox(SECOND_ELEMENT).should('not.be.disabled');
-  textbox(SECOND_ELEMENT).children()
+  textbox(positionOfElement('second')).should('not.be.disabled');
+  textbox(positionOfElement('second')).children()
     .should('not.have.css', 'color', 'rgba(0, 0, 0, 0.55)')
     .and('not.have.css', 'cursor', 'not-allowed');
 });
 
 Then('Textbox component is readOnly', () => {
   cy.wait(100, { log: DEBUG_FLAG }); // added due to animation changing
+  const borderColor = 'rgb(204, 214, 218)';
   textbox().should('have.css', 'background-color', 'rgb(249, 250, 251)')
-    .and('have.css', 'border-bottom-color', 'rgb(204, 214, 218)')
-    .and('have.css', 'border-left-color', 'rgb(204, 214, 218)')
-    .and('have.css', 'border-right-color', 'rgb(204, 214, 218)')
-    .and('have.css', 'border-top-color', 'rgb(204, 214, 218)');
+    .and('have.css', 'border-bottom-color', borderColor)
+    .and('have.css', 'border-left-color', borderColor)
+    .and('have.css', 'border-right-color', borderColor)
+    .and('have.css', 'border-top-color', borderColor);
 });
 
 Then('Textbox multiple component is readOnly', () => {
+  const borderReadonlyColor = 'rgb(204, 214, 218)';
+  const backgroundColor = 'rgb(249, 250, 251)';
   cy.wait(100, { log: DEBUG_FLAG }); // added due to animation changing
-  textbox(FIRST_ELEMENT).should('have.css', 'background-color', 'rgb(249, 250, 251)')
-    .and('have.css', 'border-bottom-color', 'rgb(204, 214, 218)')
-    .and('have.css', 'border-left-color', 'rgb(204, 214, 218)')
-    .and('have.css', 'border-right-color', 'rgb(204, 214, 218)')
-    .and('have.css', 'border-top-color', 'rgb(204, 214, 218)');
-  textbox(SECOND_ELEMENT).should('have.css', 'background-color', 'rgb(249, 250, 251)')
-    .and('have.css', 'border-bottom-color', 'rgb(204, 214, 218)')
-    .and('have.css', 'border-left-color', 'rgb(204, 214, 218)')
-    .and('have.css', 'border-right-color', 'rgb(204, 214, 218)')
-    .and('have.css', 'border-top-color', 'rgb(204, 214, 218)');
+  textbox(positionOfElement('first')).should('have.css', 'background-color', backgroundColor)
+    .and('have.css', 'border-bottom-color', borderReadonlyColor)
+    .and('have.css', 'border-left-color', borderReadonlyColor)
+    .and('have.css', 'border-right-color', borderReadonlyColor)
+    .and('have.css', 'border-top-color', borderReadonlyColor);
+  textbox(positionOfElement('second')).should('have.css', 'background-color', backgroundColor)
+    .and('have.css', 'border-bottom-color', borderReadonlyColor)
+    .and('have.css', 'border-left-color', borderReadonlyColor)
+    .and('have.css', 'border-right-color', borderReadonlyColor)
+    .and('have.css', 'border-top-color', borderReadonlyColor);
 });
 
 Then('Textbox component is not readOnly', () => {
@@ -87,29 +89,31 @@ Then('Textbox component is not readOnly', () => {
 });
 
 Then('Textbox multiple component is not readOnly', () => {
+  const borderNoReadonlyColor = 'rgba(0, 0, 0, 0)';
+  const backgroundColor = 'rgb(255, 255, 255)';
   cy.wait(100, { log: DEBUG_FLAG }); // added due to animation changing
-  textbox(FIRST_ELEMENT).should('not.have.css', 'background-color', 'rgba(0, 0, 0, 0)')
-    .and('not.have.css', 'border-bottom-color', 'rgba(0, 0, 0, 0)')
-    .and('not.have.css', 'border-left-color', 'rgba(0, 0, 0, 0)')
-    .and('not.have.css', 'border-right-color', 'rgba(0, 0, 0, 0)')
-    .and('not.have.css', 'border-top-color', 'rgba(0, 0, 0, 0)')
-    .and('have.css', 'background-color', 'rgb(255, 255, 255)');
-  textbox(SECOND_ELEMENT).should('not.have.css', 'background-color', 'rgba(0, 0, 0, 0)')
-    .and('not.have.css', 'border-bottom-color', 'rgba(0, 0, 0, 0)')
-    .and('not.have.css', 'border-left-color', 'rgba(0, 0, 0, 0)')
-    .and('not.have.css', 'border-right-color', 'rgba(0, 0, 0, 0)')
-    .and('not.have.css', 'border-top-color', 'rgba(0, 0, 0, 0)')
-    .and('have.css', 'background-color', 'rgb(255, 255, 255)');
+  textbox(positionOfElement('first')).should('not.have.css', 'background-color', borderNoReadonlyColor)
+    .and('not.have.css', 'border-bottom-color', borderNoReadonlyColor)
+    .and('not.have.css', 'border-left-color', borderNoReadonlyColor)
+    .and('not.have.css', 'border-right-color', borderNoReadonlyColor)
+    .and('not.have.css', 'border-top-color', borderNoReadonlyColor)
+    .and('have.css', 'background-color', backgroundColor);
+  textbox(positionOfElement('second')).should('not.have.css', 'background-color', borderNoReadonlyColor)
+    .and('not.have.css', 'border-bottom-color', borderNoReadonlyColor)
+    .and('not.have.css', 'border-left-color', borderNoReadonlyColor)
+    .and('not.have.css', 'border-right-color', borderNoReadonlyColor)
+    .and('not.have.css', 'border-top-color', borderNoReadonlyColor)
+    .and('have.css', 'background-color', backgroundColor);
 });
 
 Then('Multiple fieldHelp on preview is set to {string}', (text) => {
-  fieldHelpPreviewByPosition(FIRST_ELEMENT).should('have.text', text);
-  fieldHelpPreviewByPosition(SECOND_ELEMENT).should('have.text', text);
+  fieldHelpPreviewByPosition(positionOfElement('first')).should('have.text', text);
+  fieldHelpPreviewByPosition(positionOfElement('second')).should('have.text', text);
 });
 
 Then('Multiple label is set to {string}', (text) => {
-  labelByPosition(FIRST_ELEMENT).should('have.text', text);
-  labelByPosition(SECOND_ELEMENT).should('have.text', text);
+  labelByPosition(positionOfElement('first')).should('have.text', text);
+  labelByPosition(positionOfElement('second')).should('have.text', text);
 });
 
 Then('Textbox component is labelInline', () => {
@@ -117,8 +121,8 @@ Then('Textbox component is labelInline', () => {
 });
 
 Then('Multiple Textbox component is labelInline', () => {
-  labelByPosition(FIRST_ELEMENT).should('have.css', TEXT_ALIGN, 'left');
-  labelByPosition(SECOND_ELEMENT).should('have.css', TEXT_ALIGN, 'left');
+  labelByPosition(positionOfElement('first')).should('have.css', TEXT_ALIGN, 'left');
+  labelByPosition(positionOfElement('second')).should('have.css', TEXT_ALIGN, 'left');
 });
 
 Then('Textbox component is not labelInline', () => {
@@ -126,13 +130,13 @@ Then('Textbox component is not labelInline', () => {
 });
 
 Then('Multiple Textbox component is not labelInline', () => {
-  labelByPosition(FIRST_ELEMENT).should('not.have.css', TEXT_ALIGN, 'left');
-  labelByPosition(SECOND_ELEMENT).should('not.have.css', TEXT_ALIGN, 'left');
+  labelByPosition(positionOfElement('first')).should('not.have.css', TEXT_ALIGN, 'left');
+  labelByPosition(positionOfElement('second')).should('not.have.css', TEXT_ALIGN, 'left');
 });
 
 Then('Multiple tooltipPreview on preview is set to {string}', (text) => {
-  tooltipPreviewByPosition(FIRST_ELEMENT).should('have.text', text);
-  tooltipPreviewByPosition(SECOND_ELEMENT).should('have.text', text);
+  tooltipPreviewByPosition(positionOfElement('first')).should('have.text', text);
+  tooltipPreviewByPosition(positionOfElement('second')).should('have.text', text);
 });
 
 Then('Textbox inputWidth is set to {string}', (width) => {
@@ -140,13 +144,13 @@ Then('Textbox inputWidth is set to {string}', (width) => {
 });
 
 Then('Multiple Textbox inputWidth is set to {string}', (width) => {
-  textbox(FIRST_ELEMENT).should('have.css', 'flex', `0 0 ${width}%`);
-  textbox(SECOND_ELEMENT).should('have.css', 'flex', `0 0 ${width}%`);
+  textbox(positionOfElement('first')).should('have.css', 'flex', `0 0 ${width}%`);
+  textbox(positionOfElement('second')).should('have.css', 'flex', `0 0 ${width}%`);
 });
 
 Then('Multiple label width is set to {string}', (width) => {
-  labelByPosition(FIRST_ELEMENT).should('have.attr', 'width', `${width}`);
-  labelByPosition(SECOND_ELEMENT).should('have.attr', 'width', `${width}`);
+  labelByPosition(positionOfElement('first')).should('have.attr', 'width', `${width}`);
+  labelByPosition(positionOfElement('second')).should('have.attr', 'width', `${width}`);
 });
 
 When('I input {string} into Textbox', (text) => {
@@ -154,15 +158,7 @@ When('I input {string} into Textbox', (text) => {
 });
 
 When('I input {string} into {string} Textbox', (text, position) => {
-  switch (position) {
-    case 'first':
-      textboxByPosition(FIRST_ELEMENT).children().clear().type(text);
-      break;
-    case 'second':
-      textboxByPosition(SECOND_ELEMENT).children().clear().type(text);
-      break;
-    default: throw new Error('There are only two textbox elements on the page');
-  }
+  textboxByPosition(positionOfElement(position)).children().clear().type(text);
 });
 
 When('I input {string} into Textbox for deprecated component', (text) => {
@@ -176,10 +172,10 @@ Then('Textbox input on preview is set to {string}', () => {
 });
 
 Then('Multiple textbox input on preview is set to {string}', () => {
-  textbox(FIRST_ELEMENT).children().invoke('text').then(((text) => {
+  textbox(positionOfElement('first')).children().invoke('text').then(((text) => {
     expect(text.trim()).to.eq(text);
   }));
-  textbox(SECOND_ELEMENT).children().invoke('text').then(((text) => {
+  textbox(positionOfElement('second')).children().invoke('text').then(((text) => {
     expect(text.trim()).to.eq(text);
   }));
 });
@@ -195,8 +191,8 @@ Then('Textbox height is {string}', (height) => {
 });
 
 Then('Multiple Textbox height is {string}', (height) => {
-  commonInputPreview(FIRST_ELEMENT).should('have.css', 'height', height);
-  commonInputPreview(SECOND_ELEMENT).should('have.css', 'height', height);
+  commonInputPreview(positionOfElement('first')).should('have.css', 'height', height);
+  commonInputPreview(positionOfElement('second')).should('have.css', 'height', height);
 });
 
 Then('Textbox width is {string}', (width) => {
@@ -204,13 +200,13 @@ Then('Textbox width is {string}', (width) => {
 });
 
 Then('Multiple Textbox width is {string}', (width) => {
-  commonInputPreview(FIRST_ELEMENT).should('have.css', 'width', width);
-  commonInputPreview(SECOND_ELEMENT).should('have.css', 'width', width);
+  commonInputPreview(positionOfElement('first')).should('have.css', 'width', width);
+  commonInputPreview(positionOfElement('second')).should('have.css', 'width', width);
 });
 
 Then('Multiple label Align on preview is {string}', (direction) => {
-  labelByPosition(FIRST_ELEMENT).should($element => expect($element).to.have.css(TEXT_ALIGN, `${direction}`));
-  labelByPosition(SECOND_ELEMENT).should($element => expect($element).to.have.css(TEXT_ALIGN, `${direction}`));
+  labelByPosition(positionOfElement('first')).should($element => expect($element).to.have.css(TEXT_ALIGN, `${direction}`));
+  labelByPosition(positionOfElement('second')).should($element => expect($element).to.have.css(TEXT_ALIGN, `${direction}`));
 });
 
 Then('I click on icon inside of Textbox', () => {